content_engine/requirements.txt
# Content Engine Dependencies

# Core
pyyaml>=6.0
pydantic>=2.0

# Web (Milestone 3)
fastapi>=0.100
uvicorn>=0.22
jinja2>=3.1
python-multipart>=0.0.6

# Testing
pytest>=7.0
pytest-asyncio>=0.21
httpx>=0.24
